+++ packages/live-update/src/web.ts
@@ -7,14 +7,17 @@ import type {
   GetBundleResult,
   GetBundlesResult,
   GetChannelResult,
+  GetCurrentBundleResult,
   GetCustomIdResult,
   GetDeviceIdResult,
+  GetNextBundleResult,
   GetVersionCodeResult,
   GetVersionNameResult,
   LiveUpdatePlugin,
   SetBundleOptions,
   SetChannelOptions,
   SetCustomIdOptions,
+  SetNextBundleOptions,
   SyncResult,
 } from './definitions';
 
@@ -43,11 +46,19 @@ export class LiveUpdateWeb extends WebPlugin implements LiveUpdatePlugin {
     this.throwUnimplementedError();
   }
 
+  public async getCurrentBundle(): Promise<GetCurrentBundleResult> {
+    this.throwUnimplementedError();
+  }
+
   public async getCustomId(): Promise<GetCustomIdResult> {
     this.throwUnimplementedError();
   }
 
   public async getDeviceId(): Promise<GetDeviceIdResult> {
+    this.throwUnimplementedError();
+  }
+
+  public async getNextBundle(): Promise<GetNextBundleResult> {
     this.throwUnimplementedError();
   }
 
@@ -80,6 +91,10 @@ export class LiveUpdateWeb extends WebPlugin implements LiveUpdatePlugin {
   }
 
   public async setCustomId(_options: SetCustomIdOptions): Promise<void> {
+    this.throwUnimplementedError();
+  }
+
+  public async setNextBundle(_options: SetNextBundleOptions): Promise<void> {
     this.throwUnimplementedError();
   }